+++ MathParserWPF/Model/MathOperations/DivisionCommand.cs
@@ -10,12 +10,12 @@ namespace MathParserWPF.Model.MathOperations
         //    return true;
         //}
 
-        public double Execute(double a, double b)
+        public decimal Execute(decimal a, decimal b)
         {
             if (b == 0) throw new ComputingException("Деление на 0 невозможно.");
             try
             {
-                double res = a / b;
+                decimal res = a / b;
                 return res;
             }
             catch (OverflowException)
